Lesson 4 | Fixes

+++ homeworks/nodeJS/Lesson4/backend/src/controllers/brews.controller.js
@@ -8,24 +8,73 @@ export class BrewsController {
 
     list(_req, res) {
         const queryParams = _req.query;
-        return res.json(this.brewService.list(queryParams));
+        console.log("Params ", queryParams);
+        const result = this.brewService.list(queryParams);
+
+        if (!result.length) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        return res.json(result);
     }
 
     item(req, res) {
-        return res.json(this.brewService.getOne(req.params.id))
+        const id = req?.params?.id;
+
+        if (!id) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        const item = this.brewService.getOne(id);
+
+        if (!item) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        return res.json(item);
     }
 
     create(req, res) {
         console.log("START CREATING ", req.body);
+
+        if (!req.body) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
         return res.status(201).json(this.brewService.create(req.body));
     }
 
     update(req, res) {
-        return res.json(this.brewService.update(req.params.id, req.body));
+        const id = req?.params?.id;
+
+        if (!id) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        const result = this.brewService.update(req.params.id, req.body)
+
+        console.log("RESULT ", result);
+
+        if (!result) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        return res.json(result);
     }
 
     delete(req, res) {
-        this.brewService.delete(req.params.id);
+        const id = req?.params?.id;
+
+        if (!id) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
+        const item = this.brewService.delete(req.params.id);
+
+        if (!item) {
+            return res.status(404).json({ error: 'Not found' });
+        }
+
         res.status(204).end();
     };
 }
